renamed monitor to signal
Closes #147
