+++ signal/tcp_server.go
@@ -1,4 +1,4 @@
-package monitor
+package signal
 
 import (
 	"errors"
